# Some performance improvements and more idiomatic code in some places