--- a/src/camera.rs
+++ b/src/camera.rs
@@ -3,6 +3,7 @@ use crate::{
     vec3::{Point3, Vec3},
 };
 
+#[derive(Clone, Copy, Debug)]
 pub struct Camera {
     origin: Point3,
     lower_left_corner: Point3,