--- a/src/hittable.rs
+++ b/src/hittable.rs
@@ -16,23 +16,22 @@ pub trait Hittable {
 }
 
 impl HitRecord {
-    pub fn with_values(p: Point3, t: f64, material: Rc<dyn Material>) -> HitRecord {
+    pub fn new(p: Point3, t: f64, material: Option<Rc<dyn Material>>) -> HitRecord {
         HitRecord {
             p,
             normal: p,
-            material: Some(material),
+            material,
             t,
             front_face: false,
         }
     }
 
-    #[inline]
-    pub fn set_face_normal(&mut self, ray: &Ray, outward_normal: &Vec3) -> () {
-        self.front_face = Vec3::dot(&ray.direction, outward_normal) < 0.0;
-        if self.front_face {
-            self.normal = outward_normal.clone();
+    pub fn set_face_normal(&mut self, ray: &Ray, outward_normal: Vec3) {
+        self.front_face = ray.direction.dot(outward_normal) < 0.0;
+        self.normal = if self.front_face {
+            outward_normal
         } else {
-            self.normal = -outward_normal.clone();
-        }
+            -outward_normal
+        };
     }
 }