--- a/src/hittable.rs
+++ b/src/hittable.rs
@@ -1,12 +1,11 @@
 use crate::material::Material;
 use crate::ray::Ray;
 use crate::vec3::{Point3, Vec3};
-use std::rc::Rc;
 
-pub struct HitRecord {
+pub struct HitRecord<'world> {
     pub p: Point3,
     pub normal: Vec3,
-    pub material: Option<Rc<dyn Material>>,
+    pub material: Option<&'world dyn Material>,
     pub t: f64,
     pub front_face: bool,
 }
@@ -15,8 +14,8 @@ pub trait Hittable {
     fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord>;
 }
 
-impl HitRecord {
-    pub fn new(p: Point3, t: f64, material: Option<Rc<dyn Material>>) -> HitRecord {
+impl<'world> HitRecord<'world> {
+    pub fn new(p: Point3, t: f64, material: Option<&'world dyn Material>) -> Self {
         HitRecord {
             p,
             normal: p,