--- a/src/hittable_list.rs
+++ b/src/hittable_list.rs
@@ -1,9 +1,8 @@
 use crate::hittable::{HitRecord, Hittable};
 use crate::ray::Ray;
-use std::rc::Rc;
 
 pub struct HittableList {
-    objects: Vec<Rc<dyn Hittable>>,
+    objects: Vec<Box<dyn Hittable>>,
 }
 
 impl HittableList {
@@ -13,7 +12,7 @@ impl HittableList {
         }
     }
 
-    pub fn from_hittable(object: Rc<dyn Hittable>) -> HittableList {
+    pub fn from_hittable(object: impl Hittable + 'static) -> HittableList {
         let mut list = HittableList {
             objects: Vec::new(),
         };
@@ -21,12 +20,12 @@ impl HittableList {
         list
     }
 
-    pub fn clear(&mut self) -> () {
+    pub fn clear(&mut self) {
         self.objects.clear()
     }
 
-    pub fn add(&mut self, object: Rc<dyn Hittable>) -> () {
-        self.objects.push(object)
+    pub fn add(&mut self, object: impl Hittable + 'static) {
+        self.objects.push(Box::new(object))
     }
 }
 
@@ -36,12 +35,9 @@ impl Hittable for HittableList {
         let mut closest_so_far = t_max;
 
         for object in &self.objects {
-            match object.hit(ray, t_min, closest_so_far) {
-                Some(rec) => {
-                    closest_so_far = rec.t;
-                    temp_rec.replace(rec);
-                }
-                None => {}
+            if let Some(rec) = object.hit(ray, t_min, closest_so_far) {
+                closest_so_far = rec.t;
+                temp_rec.replace(rec);
             }
         }