--- a/src/main.rs
+++ b/src/main.rs
@@ -5,7 +5,6 @@ mod hittable_list;
 mod material;
 mod ray;
 mod sphere;
-mod utils;
 mod vec3;
 
 use camera::Camera;
@@ -13,83 +12,70 @@ use color::write_color;
 use hittable::Hittable;
 use hittable_list::HittableList;
 use material::{Dielectric, Lambertian, Material, Metal};
+use rand::Rng;
 use ray::Ray;
 use sphere::Sphere;
-use std::io::{self, Write};
 use std::rc::Rc;
-use utils::{random_f64, random_f64_minmax};
 use vec3::{Color, Point3, Vec3};
 
-fn random_scene() -> HittableList {
+fn random_scene(rng: &mut impl rand::Rng) -> HittableList {
     let mut world = HittableList::new();
 
-    let ground_material = Rc::new(Lambertian::with_values(&Color::with_values(0.5, 0.5, 0.5)));
-    world.add(Rc::new(Sphere::with_values(
-        Point3::with_values(0.0, -1000.0, 0.0),
+    let ground_material = Rc::new(Lambertian::new(Color::new(0.5, 0.5, 0.5)));
+    world.add(Sphere::new(
+        Point3::new(0.0, -1000.0, 0.0),
         1000.0,
         ground_material,
-    )));
+    ));
 
     for a in -11..11 {
         for b in -11..11 {
-            let choose_mat = random_f64();
-            let center = Point3::with_values(
-                a as f64 + 0.9 * random_f64(),
+            let choose_mat: f64 = rng.gen();
+            let center = Point3::new(
+                a as f64 + 0.9 * rng.gen::<f64>(),
                 0.2,
-                b as f64 + 0.9 * random_f64(),
+                b as f64 + 0.9 * rng.gen::<f64>(),
             );
 
-            if (center - Point3::with_values(4.0, 0.2, 0.0)).length() > 0.9 {
+            if (center - Point3::new(4.0, 0.2, 0.0)).length() > 0.9 {
                 let sphere_material: Rc<dyn Material>;
 
                 if choose_mat < 0.8 {
                     // diffuse
-                    let albedo = Vec3::vec3_random() * Vec3::vec3_random();
-                    sphere_material = Rc::new(Lambertian::with_values(&albedo));
-                    world.add(Rc::new(Sphere::with_values(center, 0.2, sphere_material)));
+                    let albedo = Vec3::vec3_random(rng) * Vec3::vec3_random(rng);
+                    sphere_material = Rc::new(Lambertian::new(albedo));
+                    world.add(Sphere::new(center, 0.2, sphere_material));
                 } else if choose_mat < 0.95 {
                     // metal
-                    let albedo = Vec3::vec3_random_minmax(0.5, 1.0);
-                    let fuzz = random_f64_minmax(0.0, 0.5);
-                    sphere_material = Rc::new(Metal::with_values(&albedo, fuzz));
-                    world.add(Rc::new(Sphere::with_values(center, 0.2, sphere_material)));
+                    let albedo = Vec3::vec3_random_range(rng, 0.5..1.0);
+                    let fuzz = rng.gen_range(0.0..0.5);
+                    sphere_material = Rc::new(Metal::new(albedo, fuzz));
+                    world.add(Sphere::new(center, 0.2, sphere_material));
                 } else {
                     // glass
-                    sphere_material = Rc::new(Dielectric::with_values(1.5));
-                    world.add(Rc::new(Sphere::with_values(center, 0.2, sphere_material)));
+                    sphere_material = Rc::new(Dielectric::new(1.5));
+                    world.add(Sphere::new(center, 0.2, sphere_material));
                 }
             }
         }
     }
 
-    let material1 = Rc::new(Dielectric::with_values(1.5));
-    world.add(Rc::new(Sphere::with_values(
-        Point3::with_values(0.0, 1.0, 0.0),
-        1.0,
-        material1,
-    )));
+    let material1 = Rc::new(Dielectric::new(1.5));
+    world.add(Sphere::new(Point3::new(0.0, 1.0, 0.0), 1.0, material1));
 
-    let material2 = Rc::new(Lambertian::with_values(&Color::with_values(0.4, 0.2, 0.1)));
-    world.add(Rc::new(Sphere::with_values(
-        Point3::with_values(-4.0, 1.0, 0.0),
-        1.0,
-        material2,
-    )));
+    let material2 = Rc::new(Lambertian::new(Color::new(0.4, 0.2, 0.1)));
+    world.add(Sphere::new(Point3::new(-4.0, 1.0, 0.0), 1.0, material2));
 
-    let material3 = Rc::new(Metal::with_values(&Color::with_values(0.7, 0.6, 0.5), 0.0));
-    world.add(Rc::new(Sphere::with_values(
-        Point3::with_values(4.0, 1.0, 0.0),
-        1.0,
-        material3,
-    )));
+    let material3 = Rc::new(Metal::new(Color::new(0.7, 0.6, 0.5), 0.0));
+    world.add(Sphere::new(Point3::new(4.0, 1.0, 0.0), 1.0, material3));
 
     world
 }
 
-fn hit_sphere(center: &Point3, radius: f64, ray: &Ray) -> f64 {
+fn hit_sphere(center: Point3, radius: f64, ray: &Ray) -> f64 {
     let oc = ray.origin - center;
     let a = ray.direction.length_squared();
-    let half_b = Vec3::dot(&oc, &ray.direction);
+    let half_b = oc.dot(ray.direction);
     let c = oc.length_squared() - radius * radius;
     let discriminant = half_b * half_b - a * c;
     if discriminant < 0.0 {
@@ -99,9 +85,9 @@ fn hit_sphere(center: &Point3, radius: f64, ray: &Ray) -> f64 {
     }
 }
 
-fn ray_color(ray: &Ray, world: &dyn Hittable, depth: i32) -> Color {
+fn ray_color(ray: &Ray, world: &impl Hittable, depth: i32) -> Color {
     if depth <= 0 {
-        return Color::new();
+        return Color::default();
     }
 
     if let Some(hit_record) = world.hit(ray, 0.001, f64::INFINITY) {
@@ -109,21 +95,21 @@ fn ray_color(ray: &Ray, world: &dyn Hittable, depth: i32) -> Color {
             .material
             .as_ref()
             .unwrap()
-            .scatter(ray, &hit_record)
+            .scatter(&ray, &hit_record)
         {
             return attenuation * ray_color(&scattered, world, depth - 1);
         } else {
-            return Color::new();
+            return Color::default();
         }
     }
-    let unit_direction: Vec3 = Vec3::unit_vector(ray.direction);
+    let unit_direction = ray.direction.normalize();
     let t = 0.5 * (unit_direction.y + 1.0);
-    (1.0 - t) * Color::with_values(1.0, 1.0, 1.0) + t * Color::with_values(0.5, 0.7, 1.0)
+    (1.0 - t) * Color::new(1.0, 1.0, 1.0) + t * Color::new(0.5, 0.7, 1.0)
 }
 
 fn main() {
     // TODO: error handling
-    let stdout = io::stdout();
+    let stdout = std::io::stdout();
     let mut handle = stdout.lock();
 
     // Image
@@ -133,20 +119,23 @@ fn main() {
     const SAMPLES_PER_PIXEL: i32 = 500;
     const MAX_DEPTH: i32 = 50;
 
+    // Cache thread rng
+    let mut rng = rand::thread_rng();
+
     // World
-    let world: HittableList = random_scene();
+    let world: HittableList = random_scene(&mut rng);
 
     // Camera
-    let lookfrom = Point3::with_values(13.0, 2.0, 3.0);
-    let lookat = Point3::with_values(0.0, 0.0, 0.0);
-    let vup = Vec3::with_values(0.0, 1.0, 0.0);
+    let lookfrom = Point3::new(13.0, 2.0, 3.0);
+    let lookat = Point3::new(0.0, 0.0, 0.0);
+    let vup = Vec3::new(0.0, 1.0, 0.0);
     let dist_to_focus = 10.0;
     let aperture = 0.1;
 
     let cam: Camera = Camera::new(
-        &lookfrom,
-        &lookat,
-        &vup,
+        lookfrom,
+        lookat,
+        vup,
         20.0,
         ASPECT_RATIO,
         aperture,
@@ -157,13 +146,13 @@ fn main() {
 
     println!("P3\n{} {}\n255", IMAGE_WIDTH, IMAGE_HEIGHT);
 
-    for j in (0..=IMAGE_HEIGHT - 1).into_iter().rev() {
+    for j in (0..IMAGE_HEIGHT).rev().take(5) {
         eprintln!("Scanlines remaining {}", j);
         for i in 0..IMAGE_WIDTH {
-            let mut pixel_color = Color::new();
-            for s in 0..SAMPLES_PER_PIXEL {
-                let u = (i as f64 + random_f64()) / (IMAGE_WIDTH - 1) as f64;
-                let v = (j as f64 + random_f64()) / (IMAGE_HEIGHT - 1) as f64;
+            let mut pixel_color = Color::default();
+            for _ in 0..SAMPLES_PER_PIXEL {
+                let u = (i as f64 + rng.gen::<f64>()) / (IMAGE_WIDTH - 1) as f64;
+                let v = (j as f64 + rng.gen::<f64>()) / (IMAGE_HEIGHT - 1) as f64;
                 let ray = cam.get_ray(u, v);
                 pixel_color += ray_color(&ray, &world, MAX_DEPTH);
             }