--- a/src/utils.rs
+++ b/src/utils.rs
@@ -1,4 +1,0 @@
-#[inline]
-pub fn degrees_to_radians(deg: f64) -> f64 {
-    deg * std::f64::consts::PI / 180.0
-}