# Introduction
Hi! I really liked your blog post and took a look at the code after reading it.
These are some ideas I had for making the code more idiomatic / faster.

I'd be happy to have a discussion about the things I changed and why I think they make sense or why you don't think they do.

# Benchmarks
Here are some `time`s on my machine (8/16 cores/threads) with `SAMPLES_PER_PIXEL=50`:

## f290e44 - current master:
```
...
Scanlines remaining 3
Scanlines remaining 2
Scanlines remaining 1
Scanlines remaining 0
Done
cargo run --release > /dev/null  289.71s user 0.34s system 99% cpu 4:50.22 total
```

## 033ad93 - pre-parallelization:
```
...
Scanlines remaining 3
Scanlines remaining 2
Scanlines remaining 1
Scanlines remaining 0
Done
cargo run --release > /dev/null  273.98s user 0.41s system 99% cpu 4:34.75 total
```

## 8e19696 - post-parallelization:
```
Calculated 960000/960000 pixels (100.0%)
Writing pixel 960000/960000 (100.0%)
Done
cargo run --release > /dev/null  730.92s user 0.77s system 1563% cpu 46.795 total
```