@JDuchniewicz 2021-03-01
Thanks for the PR! It looks pretty solid. Currently I am on holidays but I will merge it as soon as I have time to read into it in details 😁 

BTW. Linked it on the reddit post in [comments](https://www.reddit.com/r/rust/comments/lukgyi/c_to_rust_introduction_with_practical_raytracing/gpb0qgk?utm_source=share&utm_medium=web2x&context=3).