@JDuchniewicz 2021-03-09
Revert done. The `inline` specifiers were the only thing I think.