@soruh 2021-03-09
Okay, so i have looked into this and seems like it's not as easy as I though...
(Note that the following benchmarks are very unscientific and the random start scene might have a significant impact. I ran all of these a few times an the times looked relatively stable)

Baseline speed is around 47s for 50 Samples for me.
turning all `#[inline]` annotations into `#[inline(always)]` makes the code slightly slower at 48s
and removing all `#[inline]` annotations actually slows us down even more to 50s.

It looks like `#[inline]` also inlines across code-gen units. Since code optimized for speed is usually compiled with only
one codegen unit that shouldn't be relevant for us though:

After tuning the Cargo.toml a bit:
```toml
[profile.release]
codegen-units = 1
lto = true
```
the baseline stays the same at 47s, but the version without the `#[inline]` annotations goes down to 45s