@JDuchniewicz 2021-03-09
Looks good to me. In a more complicated rendering engine it would be beneficial to see if these `#[inline]`s can speedup the execution.